+++ engine/musicbrainz_label_onebox.rb
@@ -13,7 +13,7 @@ module Onebox
       private
 
       def url
-        "https://#{match[:domain]}/ws/2/#{@@entity}/#{match[:mbid]}?fmt=json&inc=url-rels"
+        "https://#{match[:domain]}/ws/2/#{@@entity}/#{match[:mbid]}?fmt=json&inc=url-rels+genres"
       end
 
       def match
@@ -32,6 +32,7 @@ module Onebox
         disambiguation
         area
         life_span
+        genres
         add_critiquebrainz_link
         image "logo"
         wikidata