+++ engine/musicbrainz_releasegroup_onebox.rb
@@ -13,7 +13,7 @@ module Onebox
       private
 
       def url
-        "https://#{match[:domain]}/ws/2/#{@@entity}/#{match[:mbid]}?fmt=json&inc=artist-credits"
+        "https://#{match[:domain]}/ws/2/#{@@entity}/#{match[:mbid]}?fmt=json&inc=artist-credits+genres"
       end
 
       def match
@@ -33,6 +33,7 @@ module Onebox
 
         artist_credits
         disambiguation
+        genres
         add_critiquebrainz_link
         caa_image
         wikidata