+++ CHANGES.md
@@ -2,6 +2,7 @@
 - CritiqueBrainz links for recordings
 - Show Wikipedia link for all entities, if available
 - Removed AcousticBrainz link (#18)
+- Show genres for artists, labels, release groups, releases and recordings
 
 # 0.8.0 - 2022-12-12
 - Show link to release group on CritiqueBrainz in release onebox (#17)